Major Updates to Customer
Added a customer controller. I decided that it was better to be consistient and have a dedicated customer controller. Also added a customer version of a project component view that even has a dynamic animated progress bar which will reflect the overall project progress. Finally decided to remove the register and login links at the top of the nav bar since they were redundant.