Major Changes to Scope MVC
Changes to the scope model to include a numerical representation of scope. This is to allow for a percentage to be shown in the customer view of a project. Also fixed up the view of the scope for employees and customers. Also fixed the UI text size being too small.