Initial stages of redesign and rewrite. New theme.
This project will most likely migrate to a different repo in the near future to remove all branding. The changes being pushed are the result of initial UI redesign and some minor controller logic updates. There is going to be a migration of more bussiness logic being handled within domain-like classes which will help to simplify controller/view code.